add test account.

+++ routes/echelon/profile.js
@@ -93,6 +93,11 @@ router.get('/user/:user_id', async (req, res) => {
 //access: private
 router.delete('/me', auth, async (req, res)=> {
     try {
+        //can't remove test account
+        const testAccount = User.findById({_id: '5d9857819794740017eb6df3'})
+        if (testAccount){
+            return res.json({ msg: 'You can not delete test account.'})
+        }
         //remove user's posts 
         await Post.deleteMany({user: req.user.id})
         //remove profile and user
